refactor(KERN): move textarea folder
