+++ app/components/formElements/__test__/Textarea.test.tsx
@@ -1,7 +1,5 @@
 import { render, screen, fireEvent } from "@testing-library/react";
-import Textarea, {
-  TEXT_AREA_ROWS,
-} from "~/components/kern/formElements/Textarea";
+import Textarea, { TEXT_AREA_ROWS } from "~/components/formElements/Textarea";
 import { TEXTAREA_CHAR_LIMIT } from "~/services/validation/inputlimits";
 
 const getInputProps = vi.fn();
