+++ app/components/kern/KernOpenQuestion.tsx
@@ -2,7 +2,7 @@ import { type BasicSurveyQuestion } from "posthog-js";
 import { useId, type Dispatch, type SetStateAction } from "react";
 import { TEXTAREA_CHAR_LIMIT } from "~/services/validation/inputlimits";
 import { questionToAnswerId } from "../../services/analytics/surveys/questionToAnswerId";
-import { TEXT_AREA_ROWS } from "./formElements/Textarea";
+import { TEXT_AREA_ROWS } from "../formElements/Textarea";
 
 export type SurveyResponses = Record<string, string | string[]>;
