+++ stories/form/Textarea.stories.tsx
@@ -4,7 +4,7 @@ import { Grid } from "~/components/layout/grid/Grid";
 import { GridSection } from "~/components/layout/grid/GridSection";
 import { GridItem } from "~/components/layout/grid/GridItem";
 import { reactRouterFormContext } from "~/../.storybook/reactRouterFormContext";
-import KernTextarea from "~/components/kern/formElements/Textarea";
+import KernTextarea from "~/components/formElements/Textarea";
 
 const meta = {
   title: "form/Textarea",
